fix add

+++ app/Http/Controllers/UserController.php
@@ -91,6 +91,9 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->id == $id) {
+            return redirect()->route('users.index');
+        }
         $user = User::findOrFail($id);
         $user->delete();
         return redirect()->route('users.index');
